src/desktop/requirements.txt
websocket-client==1.3.3
pyautogui==0.9.53
qrcode==7.3.1
pillow==9.0.1